Update helper library
